x64dbgBccPlugin
Sample plugin for x64dbg

Sample Bcc v1 plugin by ThunderCls - 2016 Blog:
http://reversec0de.wordpress.com

This is a sample plugin for the x64dbg x86/x64 debugger by @mrexodia
using Borland/Embarcadero C++

Usage:

The project is a sample of an x64dbg plugin using the RAD Studio XE10
Berlin. For using this just add the code of your plugin and everything
else is already set for compiling without issues.

Installation:

Copy *.dp32 to x32 plugins directory of x64dbg
Look under the "Plugins" menu in the main x64dbg window or in the
secondary menu in the Disasm window as well